MainViewModel.cs декомпозирован, добавлены классы управляющие деревом назначений и МЕР-категориями
Добавлена возможность не загружать из шаблона новые расходники
