+++ src/RevitUnmodelingMep/assets/Localization/Language.en-US.xaml
@@ -18,7 +18,8 @@
     <system:String x:Key="MainWindow.ButtonExport">Export</system:String>
     <system:String x:Key="MainWindow.ButtonRefresh">Refresh</system:String>
     <system:String x:Key="MainWindow.ButtonDefault">Reset</system:String>
-    <system:String x:Key="MainWindow.RefreshConfirmMessage">Refreshing will reset all template consumables to the default view and add new templates. Continue?</system:String>
+    <system:String x:Key="MainWindow.RefreshConfirmMessage">Refreshing will reset all template consumables to the default view. Continue?</system:String>
+    <system:String x:Key="MainWindow.AddMissingTemplateConsumablesConfirmMessage">The template contains consumable items that do not exist in the project. Add them?</system:String>
     <system:String x:Key="MainWindow.RefreshPlaceholderMessage">Refresh is not implemented yet.</system:String>
     <system:String x:Key="MainWindow.ResetConfirmMessage">Are you sure you want to reset to default values?</system:String>
     <system:String x:Key="MainWindow.TemplateDifferenceText">Does not match the template. Refresh or rename.</system:String>
