+++ src/RevitUnmodelingMep/assets/Localization/Language.ru-RU.xaml
@@ -18,7 +18,8 @@
     <system:String x:Key="MainWindow.ButtonExport">Экспорт</system:String>
     <system:String x:Key="MainWindow.ButtonRefresh">Обновить</system:String>
     <system:String x:Key="MainWindow.ButtonDefault">Сброс</system:String>
-    <system:String x:Key="MainWindow.RefreshConfirmMessage">При обновлении все шаблонные расходники будут сброшены к виду по умолчанию, а новые шаблоны будут добавлены. Продолжить?</system:String>
+    <system:String x:Key="MainWindow.RefreshConfirmMessage">При обновлении все шаблонные расходники будут сброшены к виду по умолчанию. Продолжить?</system:String>
+    <system:String x:Key="MainWindow.AddMissingTemplateConsumablesConfirmMessage">В шаблоне обнаружены расходные элементы, не существующие в проекте. Добавить их?</system:String>
     <system:String x:Key="MainWindow.RefreshPlaceholderMessage">Обновление пока не реализовано.</system:String>
     <system:String x:Key="MainWindow.ResetConfirmMessage">Вы уверены что хотите сбросить до значений по умолчанию?</system:String>
     <system:String x:Key="MainWindow.TemplateDifferenceText">Не соответствует шаблону. Выполните обновление или переименуйте.</system:String>
